# web/static/openapi.yaml
openapi: 3.0.3
info:
  title: freehire public API
  version: 1.1.0
  description: |
    Public, read-only freehire API: search IT jobs, read the live filter
    vocabulary, inspect job and company details, and find similar jobs. Every
    operation here is unauthenticated. The authenticated tracking surface
    (saved jobs, application stages, CV tooling) is deliberately excluded — this
    schema is the integration contract for search, and is also what the freehire
    custom GPT imports as an Action.

    ## Filter grammar

    Facet parameters are **repeatable**, and the same grammar applies to all of
    them:

    * `?skills=go&skills=kubernetes` — values of one facet are **ORed**: match
      either.
    * `?skills=go&skills=kubernetes&skills_mode=and` — `<facet>_mode=and` flips
      that facet to **AND**: match both. Not available for geography (see below).
    * `?source_exclude=adzuna` — `<facet>_exclude` removes matches. **Every**
      string facet in this schema supports its `_exclude` twin, whether or not it
      is listed below; only the most commonly used ones are declared explicitly.
    * Different facets **AND** together: `?work_mode=remote&seniority=senior` is
      remote *and* senior.

    ## Geography is one OR-group

    `regions`, `countries` and `cities` name a single concept — *where* — so
    their values join into **one OR-group** instead of intersecting.
    `?countries=gb&cities=London` therefore returns *United Kingdom **or**
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security: []
...
paths:
  /jobs/search:
    get:
      operationId: searchJobs
      tags:
        - Jobs
      summary: Search open jobs
      description: |
        Full-text and faceted search over open postings. Descriptions are the
        search index's truncated preview. Use `agentSearchJobs` when you need the
        full verbatim body of every hit, or `getJob` for the one posting the user
        picked.

        Closed postings are never in the index, so they never appear here.
      x-openai-isConsequential: false
      parameters:
        - $ref: "#/components/parameters/Q"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Order"
        - $ref: "#/components/parameters/Regions"
        - $ref: "#/components/parameters/Countries"
        - $ref: "#/components/parameters/Cities"
        - $ref: "#/components/parameters/WorkMode"
        - $ref: "#/components/parameters/Category"
        - $ref: "#/components/parameters/Role"
        - $ref: "#/components/parameters/Seniority"
        - $ref: "#/components/parameters/Skills"
        - $ref: "#/components/parameters/SkillsMode"
        - $ref: "#/components/parameters/IsTech"
        - $ref: "#/components/parameters/AIArchetype"
        - $ref: "#/components/parameters/Collections"
        - $ref: "#/components/parameters/Reality"
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/CompanySlug"
        - $ref: "#/components/parameters/EmploymentType"
        - $ref: "#/components/parameters/Relocation"
        - $ref: "#/components/parameters/EnglishLevel"
        - $ref: "#/components/parameters/EducationLevel"
        - $ref: "#/components/parameters/PostingLanguage"
        - $ref: "#/components/parameters/Domains"
        - $ref: "#/components/parameters/CompanyType"
        - $ref: "#/components/parameters/CompanySize"
        - $ref: "#/components/parameters/SalaryCurrency"
        - $ref: "#/components/parameters/SalaryPeriod"
        - $ref: "#/components/parameters/VisaSponsorship"
        - $ref: "#/components/parameters/SalaryMin"
        - $ref: "#/components/parameters/SalaryMax"
        - $ref: "#/components/parameters/ExperienceYearsMin"
        - $ref: "#/components/parameters/PostedWithinDays"
        - $ref: "#/components/parameters/RegionsExclude"
        - $ref: "#/components/parameters/CountriesExclude"
        - $ref: "#/components/parameters/WorkModeExclude"
        - $ref: "#/components/parameters/SkillsExclude"
        - $ref: "#/components/parameters/SourceExclude"
        - $ref: "#/components/parameters/CompanySlugExclude"
      responses:
        "200":
          description: Matching jobs with pagination metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListEnvelope"
        default:
          $ref: "#/components/responses/Error"
  /agent/jobs/search:
    get:
      operationId: agentSearchJobs
      tags:
        - Jobs
      summary: Search open jobs for agents (full descriptions)
      description: |
        Identical query surface to `searchJobs`, but every result carries the
        job's **full** description, read verbatim from the database rather than
        the search index's preview, in the format selected by
        `description_format`.

        Responses are several times larger per hit than `searchJobs`, so prefer a
        small `limit`.
      x-openai-isConsequential: false
      parameters:
        - $ref: "#/components/parameters/Q"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Order"
        - $ref: "#/components/parameters/DescriptionFormat"
        - $ref: "#/components/parameters/Regions"
        - $ref: "#/components/parameters/Countries"
        - $ref: "#/components/parameters/Cities"
        - $ref: "#/components/parameters/WorkMode"
        - $ref: "#/components/parameters/Category"
        - $ref: "#/components/parameters/Role"
        - $ref: "#/components/parameters/Seniority"
        - $ref: "#/components/parameters/Skills"
        - $ref: "#/components/parameters/SkillsMode"
        - $ref: "#/components/parameters/IsTech"
        - $ref: "#/components/parameters/AIArchetype"
        - $ref: "#/components/parameters/Collections"
        - $ref: "#/components/parameters/Reality"
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/CompanySlug"
        - $ref: "#/components/parameters/EmploymentType"
        - $ref: "#/components/parameters/Relocation"
        - $ref: "#/components/parameters/EnglishLevel"
        - $ref: "#/components/parameters/EducationLevel"
        - $ref: "#/components/parameters/PostingLanguage"
        - $ref: "#/components/parameters/Domains"
        - $ref: "#/components/parameters/CompanyType"
        - $ref: "#/components/parameters/CompanySize"
        - $ref: "#/components/parameters/SalaryCurrency"
        - $ref: "#/components/parameters/SalaryPeriod"
        - $ref: "#/components/parameters/VisaSponsorship"
        - $ref: "#/components/parameters/SalaryMin"
        - $ref: "#/components/parameters/SalaryMax"
        - $ref: "#/components/parameters/ExperienceYearsMin"
        - $ref: "#/components/parameters/PostedWithinDays"
        - $ref: "#/components/parameters/RegionsExclude"
        - $ref: "#/components/parameters/CountriesExclude"
        - $ref: "#/components/parameters/WorkModeExclude"
        - $ref: "#/components/parameters/SkillsExclude"
        - $ref: "#/components/parameters/SourceExclude"
        - $ref: "#/components/parameters/CompanySlugExclude"
      responses:
        "200":
          description: Matching jobs with pagination metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListEnvelope"
        default:
          $ref: "#/components/responses/Error"
  /jobs/facets:
    get:
      operationId: getJobFacets
      tags:
        - Jobs
      summary: Get live filter vocabulary and counts
      description: |
        Return facet values, counts, canonical skill slugs, and numeric ranges
        for the postings matching the same filter grammar the search endpoints
        take. Call this before applying any uncertain filter.

        A distribution is counted per facet and the wide-valued ones (`cities`,
        `skills`) dominate, so narrow with `facets=` whenever you read only a few.
        `company_slug` has no distribution — use `searchCompanies` for the company
        typeahead.
      x-openai-isConsequential: false
      parameters:
        - $ref: "#/components/parameters/Q"
        - name: facets
          in: query
          schema:
            type: string
          description: |
            Comma-separated facet names to count, e.g. `skills,seniority`. Omit
            for all of them. An unknown name is a 400. Cannot be combined with
            `disjunctive`.
        - name: disjunctive
          in: query
          schema:
            type: boolean
            default: false
          description: |
            Count each facet under the full filter MINUS its own selection, so a
            selected facet still shows its siblings' counts — the behaviour a
            live filter sidebar wants. Costs one query per facet.
        - $ref: "#/components/parameters/Regions"
        - $ref: "#/components/parameters/Countries"
        - $ref: "#/components/parameters/Cities"
        - $ref: "#/components/parameters/WorkMode"
        - $ref: "#/components/parameters/Category"
        - $ref: "#/components/parameters/Role"
        - $ref: "#/components/parameters/Seniority"
        - $ref: "#/components/parameters/Skills"
        - $ref: "#/components/parameters/SkillsMode"
        - $ref: "#/components/parameters/IsTech"
        - $ref: "#/components/parameters/AIArchetype"
        - $ref: "#/components/parameters/Collections"
        - $ref: "#/components/parameters/Reality"
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/CompanySlug"
        - $ref: "#/components/parameters/EmploymentType"
        - $ref: "#/components/parameters/Relocation"
        - $ref: "#/components/parameters/EnglishLevel"
        - $ref: "#/components/parameters/EducationLevel"
        - $ref: "#/components/parameters/PostingLanguage"
        - $ref: "#/components/parameters/Domains"
        - $ref: "#/components/parameters/CompanyType"
        - $ref: "#/components/parameters/CompanySize"
        - $ref: "#/components/parameters/SalaryCurrency"
        - $ref: "#/components/parameters/SalaryPeriod"
        - $ref: "#/components/parameters/VisaSponsorship"
        - $ref: "#/components/parameters/SalaryMin"
        - $ref: "#/components/parameters/SalaryMax"
        - $ref: "#/components/parameters/ExperienceYearsMin"
        - $ref: "#/components/parameters/PostedWithinDays"
      responses:
        "200":
          description: Facet distribution for matching jobs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacetsEnvelope"
        default:
          $ref: "#/components/responses/Error"
...
  /jobs/{slug}/similar:
    get:
      operationId: getSimilarJobs
      tags:
        - Jobs
      summary: Find jobs similar to a job
      description: |
        Nearest neighbours from a precomputed semantic list, nearest first.
        The list is built offline, so a neighbour that has since closed is
        dropped silently — a response may hold fewer items than `limit`.
      x-openai-isConsequential: false
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: Job public_slug.
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 6
          description: Maximum similar jobs to return. Values above the maximum are clamped.
      responses:
        "200":
          description: Similar jobs. This response carries no `meta` block.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Job"
        default:
          $ref: "#/components/responses/Error"
...
components:
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  parameters:
    Q:
      name: q
      in: query
      schema:
        type: string
      description: Full-text query over title, company, and description.
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Page size. Values above the maximum are clamped, not rejected.
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: |
        Rows to skip. On the job endpoints `offset + limit` may not exceed
        10000; deeper paging is a 400.
    Sort:
      name: sort
      in: query
      schema:
        type: string
        enum:
          - created_at
          - posted_at
          - salary_min
          - salary_max
      description: |
        Sort field. Omitted, a text query sorts by relevance and an empty query
        by `posted_at` descending.
    Order:
      name: order
      in: query
      schema:
        type: string
        enum:
          - asc
          - desc
        default: desc
      description: Sort direction. Ignored without a valid `sort`.
    DescriptionFormat:
      name: description_format
      in: query
      schema:
        type: string
        enum:
          - html
          - text
          - markdown
        default: html
      description: |
        Format of the full description. `html` is the stored verbatim markup;
        any unrecognized value falls back to it rather than erroring.
    Regions:
      name: regions
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - global
            - north_america
            - latam
            - eu
            - uk
            - mena
            - africa
            - apac
            - cis
            - none
      description: |
        Macro-regions. Country codes are NOT regions — use `countries`. `none`
        is the reserved value for postings with no resolved geography. Joins the
        geography OR-group with `countries` and `cities`.
    Countries:
      name: countries
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: |
        ISO 3166-1 alpha-2, lowercase (`gb`, `de`). Matching is
        case-insensitive. Joins the geography OR-group.
    Cities:
      name: cities
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: |
        Canonical city display names ("London", "Berlin") — resolve one with
        `searchCities` first. Joins the geography OR-group.
    WorkMode:
      name: work_mode
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - remote
            - hybrid
            - onsite
      description: |
        Work format, resolved from a deterministic dictionary. Absent on a
        posting whose format could not be resolved, so this filter narrows to
        postings that stated it.
    Category:
      name: category
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: Role category slug from `getJobFacets`.
    Role:
      name: role
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: |
        Fine-grained role slug from `getJobFacets` (e.g. `backend`,
        `android_developer`, and seniority-qualified variants). Much narrower
        than `category`; the vocabulary is large and generated, so always read it
        from the facets endpoint.
    Seniority:
      name: seniority
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - intern
            - junior
            - middle
            - senior
            - lead
            - staff
            - principal
            - c_level
    Skills:
      name: skills
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: Canonical skill slugs from `getJobFacets`. Never invent one.
    SkillsMode:
      name: skills_mode
      in: query
      schema:
        type: string
        enum:
          - and
      description: |
        `and` requires a posting to carry EVERY listed skill instead of any of
        them. The same `<facet>_mode=and` switch works on any non-geography
        facet.
    IsTech:
      name: is_tech
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - tech
            - non_tech
      description: |
        Technical vs non-technical, derived deterministically from title and
        category. Absent when unknown, so the filter never guesses.
    AIArchetype:
      name: ai_archetype
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - rag_app_builder
            - agent_builder
            - cloud_ml_platform_engineer
            - ml_trainer_researcher
            - fullstack_ai_engineer
            - devops_infra_engineer
      description: AI skill-signature archetype, derived from the posting's skill set.
    Collections:
      name: collections
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: |
        Curated collection slugs of the posting's company, e.g. `yc`,
        `bigtech`, `unicorn`, `us-h1b-sponsor`. Read the live set from
        `getJobFacets`.
    Reality:
      name: reality
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - fresh
            - stale
            - likely-evergreen
      description: |
        Posting-reality class. `fresh` is recently posted and not obviously
        recycled; `stale` is old; `likely-evergreen` reads as an always-open
        pipeline ad. A large share of the catalogue is `stale`, so `?reality=fresh`
        is the cheapest quality filter available. Read the live split from
        `getJobFacets`.
    Source:
      name: source
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: |
        Origin slug of the posting — the ATS or board it was crawled from
        (`greenhouse`, `workday`, `adzuna`, …). Read the live set from
        `getJobFacets`. Pair with `source_exclude` to avoid double-counting a
        board you already query directly.
    CompanySlug:
      name: company_slug
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: |
        Company slugs from `searchCompanies`. This facet has no distribution in
        `getJobFacets` — use the company search for a typeahead.
    EmploymentType:
      name: employment_type
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - full_time
            - part_time
            - contract
            - internship
            - fellowship
    Relocation:
      name: relocation
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - not_supported
            - supported
            - required
    EnglishLevel:
      name: english_level
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - none
            - a1
            - a2
            - b1
            - b2
            - c1
            - c2
            - native
    EducationLevel:
      name: education_level
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - none
            - bachelor
            - master
            - phd
    PostingLanguage:
      name: posting_language
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: Language the posting is written in (ISO 639-1, e.g. `en`, `de`, `uk`).
    Domains:
      name: domains
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - fintech
            - crypto
            - ecommerce
            - gambling
            - gamedev
            - media
            - travel
            - healthcare
            - edtech
            - govtech
            - devtools
            - cybersecurity
            - ai
            - hrtech
            - adtech
            - proptech
            - logistics
            - mobility
            - climatetech
            - other
      description: Business domain of the hiring company, as stated by the posting.
    CompanyType:
      name: company_type
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - product
            - startup
            - outsource
            - outstaff
            - agency
            - inhouse
            - government
    CompanySize:
      name: company_size
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - 1-10
            - 11-50
            - 51-200
            - 201-500
            - 501-1000
            - 1000+
    SalaryCurrency:
      name: salary_currency
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: ISO 4217 code, e.g. `USD`, `EUR`.
    SalaryPeriod:
      name: salary_period
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - year
            - month
            - day
            - hour
    VisaSponsorship:
      name: visa_sponsorship
      in: query
      schema:
        type: boolean
      description: |
        Whether the posting states it sponsors a visa. `false` is a real,
        stated value, not "unknown" — postings that say nothing match neither
        setting.
    SalaryMin:
      name: salary_min
      in: query
      schema:
        type: integer
      description: |
        Lower bound on the posting's minimum salary, in its own
        `salary_currency` and `salary_period` — the comparison does NOT
        normalize currencies. Pair with `salary_currency` and `salary_period` for
        a meaningful range.
    SalaryMax:
      name: salary_max
      in: query
      schema:
        type: integer
      description: Upper bound on the posting's maximum salary. Same currency caveat as `salary_min`.
    ExperienceYearsMin:
      name: experience_years_min
      in: query
      schema:
        type: integer
      description: Lower bound on the years of experience the posting asks for.
    PostedWithinDays:
      name: posted_within_days
      in: query
      schema:
        type: integer
        minimum: 1
      description: Restrict to postings published within the last N days. Omit for no freshness restriction.
    RegionsExclude:
      name: regions_exclude
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: Regions to exclude. Excludes AND together, unlike included geography.
    CountriesExclude:
      name: countries_exclude
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: Countries to exclude.
    WorkModeExclude:
      name: work_mode_exclude
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: Work formats to exclude, e.g. `onsite`.
    SkillsExclude:
      name: skills_exclude
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: Skill slugs to exclude.
    SourceExclude:
      name: source_exclude
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: Source slugs to exclude.
    CompanySlugExclude:
      name: company_slug_exclude
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: Company slugs to exclude.
  schemas:
...
    JobListEnvelope:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: string